+++ src/index.ts
@@ -3,3 +3,5 @@ export {
   isGaussianSplat,
   isGaussianSplatScene,
 } from './GaussianSplatPlugin';
+
+export { getSparkRendererForScene } from './SharedSparkRendererManager';